web_dashboard/tests/requirements.txt
pytest
requests
kafka-python
pymongo
pandas 